refactor: extracting Input component from SearchHeader and HomeSearchContaienr to its own file
