+++ frontend/src/components/custom_input/CustomInputStyles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const StyledInput = styled.input`
+    width: 90%;
+    display: inline-block;
+    height: 56px;
+    font-size: 16px;
+    border: none;
+    outline: none;
+    padding-left: 64px;
+`;
+
