+++ frontend/src/pages/home/components/search_container/HomeSearchContainerStyles.ts
@@ -46,16 +46,6 @@ export const StyledSearchIcon = styled(SearchIcon)`
 `;
 
 
-export const StyledInput = styled.input`
-    width: 90%;
-    display: inline-block;
-    height: 56px;
-    font-size: 16px;
-    border: none;
-    outline: none;
-    padding-left: 64px;
-`;
-
 
 
 export const GithubIconContainer = styled.div`
