+++ frontend/src/pages/search/components/header/SearchHeaderStyles.ts
@@ -50,16 +50,6 @@ export const InputContainer = styled.div`
   }
 `;
 
-export const StyledInput = styled.input`
-  border: none;
-  padding: 12px 0;
-  padding-left: 55px;
-  width: 100%;
-  outline: none;
-  display: inline-block;
-  font-size: 16px;
-`;
-
 export const StyledSearchIcon = styled(SearchIcon)`
   position: absolute;
   top: 50%;
